+++ test/ChatWork/Endpoints/ContactsSpec.hs
@@ -1,23 +1,17 @@
 module ChatWork.Endpoints.ContactsSpec
-    ( main
-    , spec
+    ( spec
     ) where
 
 import           ChatWork.Endpoints.Contacts   (getContacts)
 import           ChatWork.Test.Client          (TestClient (..))
-import           ChatWork.Test.MockServer      (runMockServer)
 import           ChatWork.Test.MonadHttpIO     ()
 import           ChatWork.Test.Values.Contacts (contacts)
 import           Network.HTTP.Req              (responseBody)
-import           Test.Hspec                    (Spec, around_, context,
-                                                describe, hspec, it,
+import           Test.Hspec                    (Spec, context, describe, it,
                                                 shouldReturn)
 
-main :: IO ()
-main = hspec spec
-
 spec :: Spec
-spec = around_ runMockServer $ do
+spec = do
   describe "getContacts: endpoint GET /contacts" $ do
     context "correct responce" $ do
       it "should return Right contacts response body" $ do