+++ test/ChatWork/Endpoints/RoomsSpec.hs
@@ -1,8 +1,7 @@
 {-# LANGUAGE OverloadedStrings #-}
 
 module ChatWork.Endpoints.RoomsSpec
-    ( main
-    , spec
+    ( spec
     ) where
 
 import           ChatWork.Endpoints.Rooms   (createRoom, createTask, deleteRoom,
@@ -13,7 +12,6 @@ import           ChatWork.Endpoints.Rooms   (createRoom, createTask, deleteRoom,
                                              updateMembersPermission,
                                              updateRoom)
 import           ChatWork.Test.Client       (TestClient (..))
-import           ChatWork.Test.MockServer   (runMockServer)
 import           ChatWork.Test.MonadHttpIO  ()
 import           ChatWork.Test.Values.Rooms (createRoomParams, createTaskParams,
                                              file, files, getTasksParams,
@@ -24,14 +22,11 @@ import           ChatWork.Test.Values.Rooms (createRoomParams, createTaskParams,
                                              roomTasks, rooms, taskIds,
                                              updateRoomParams)
 import           Network.HTTP.Req           (responseBody)
-import           Test.Hspec                 (Spec, around_, context, describe,
-                                             hspec, it, shouldReturn)
-
-main :: IO ()
-main = hspec spec
+import           Test.Hspec                 (Spec, context, describe, it,
+                                             shouldReturn)
 
 spec :: Spec
-spec = around_ runMockServer $ do
+spec = do
   describe "getRooms: endpoint GET /rooms" $ do
     context "correct responce" $ do
       it "should return Right rooms response body" $ do