+++ test/ChatWork/Test/MockServer.hs
@@ -118,5 +118,5 @@ mockServer = run 8000 (serve api server)
 
 runMockServer :: IO () -> IO ()
 runMockServer action = do
-  tid <- forkIO mockServer
-  action `finally` killThread tid
+  _ <- forkIO mockServer
+  action